add memoryerror exception to knotfold wrapper

+++ src/predict.py
@@ -118,7 +118,6 @@ def ufold_predict(seq, path_ufold):
     # https://github.com/uci-cbcl/UFold
 
     tstart = time.time()
-
     path_ufold = Path(path_ufold)
 
     # clear memory
@@ -263,7 +262,6 @@ def knotfold_predict(seq, path_knotfold="../KnotFold"):
     # https://github.com/gongtiansu/KnotFold
 
     tstart = time.time()
-
     path_knotfold = Path(path_knotfold)
 
     # clear memory
@@ -278,19 +276,21 @@ def knotfold_predict(seq, path_knotfold="../KnotFold"):
         f.write(f">{temp_rna_name}\n{seq}\n")
 
     # predict
-    subprocess.run(["python", "KnotFold.py", "-i", path_in, "-o", ".", "--cuda"],
+    res = subprocess.run(["python", "KnotFold.py", "-i", path_in, "-o", ".", "--cuda"],
                                                                 cwd=path_knotfold)
     t = torch.cuda.get_device_properties(0).total_memory
     r = torch.cuda.memory_reserved(0)
     memory = r / t
 
     # read output
+    os.remove(path_knotfold / path_in)
+    if res.returncode != 0:
+        raise MemoryError(f"The KnotFold script could not run properly. The input sequence may be too long. Avoid sequences longer than 2000 nucleotides when using KnotFold. If the sequence is shorter, then something else in the KnotFold script may have caused this error. Look in the traceback from the KnotFold script for more information.")
     with open(path_knotfold / path_out, "r") as f:
         pred_txt = f.read()
     pairs = np.array([int(line.split(" ")[-1]) for line in pred_txt.strip().split("\n")])
     pred = pairs_to_struct(pairs)
 
-    os.remove(path_knotfold / path_in)
     os.remove(path_knotfold / path_out)
 
     ttot = time.time() - tstart
